+++ components/Points.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="grid grid-cols-3 mt-20">
+    <div class="col-span-3">
+      <div class="grid grid-cols-6">
+        <div class="col-span-5">hoge</div>
+        <div class="col-span-1">fuga</div>
+      </div>
+    </div>
+  </div>
+</template>
+<script>
+import { defineComponent } from '@nuxtjs/composition-api'
+
+export default defineComponent({
+  setup() {},
+})
+</script>
